+++ internal/reconciler/reconciler.go
@@ -42,7 +42,7 @@ func ReconcileNamespace(ctx context.Context, _client client.Client, namespace st
 	}
 	objectsToReconcile = objectsToReconcile + len(deployments.Items)
 
-	scaleReplicalist, err := ObjectStateReplicasList(finalState, deployments)
+	scaleReplicalist, err := resources.DeploymentStateReplicasList(finalState, deployments)
 	if err != nil {
 		log.Error(err, "Cannot fetch replicas of all opted-in deployments")
 		return err
@@ -61,9 +61,13 @@ func ReconcileNamespace(ctx context.Context, _client client.Client, namespace st
 	if allowed {
 
 		for i, deployment := range deployments.Items {
-			err = ReconcileDeploymentforScalingState(ctx, _client, deployment, scaleReplicalist[i])
+			log := ctrl.Log.
+				WithValues("deployment", deployment.Name).
+				WithValues("namespace", deployment.Namespace)
+
+			err := resources.ScaleDeployment(ctx, _client, deployment, scaleReplicalist[i])
 			if err != nil {
-				log.Error(err, "Could not reconcile deployment.")
+				log.Error(err, "Error scaling the deployment")
 				continue
 			}
 		}
@@ -96,76 +100,13 @@ func ReconcileNamespace(ctx context.Context, _client client.Client, namespace st
 	return nil
 }
 
-func ObjectStateReplicas(state states.State, deployment v1.Deployment, optIn bool) (sr.StateReplica, error) {
-	log := ctrl.Log.
-		WithValues("deployment", deployment.Name).
-		WithValues("namespace", deployment.Namespace)
-	stateReplicas, err := sr.NewStateReplicasFromAnnotations(deployment.GetAnnotations())
-	if err != nil {
-		log.WithValues("deployment", deployment.Name).
-			WithValues("namespace", deployment.Namespace).
-			Error(err, "Cannot calculate state replicas. Please check deployment annotations. Continuing.")
-		return sr.StateReplica{}, err
-	}
-	// Now we have all the state settings, we can set the replicas for the deployment accordingly
-	if !optIn {
-		// the deployment opted out. We need to set back to default.
-		log.Info("The deployment opted out. Will scale back to default")
-		state.Name = c.DefaultReplicaAnnotation
-	}
-	stateReplica, err := stateReplicas.GetState(state.Name)
-	if err != nil {
-		// TODO here we should do priority filtering, and go down one level of priority to find the lowest set one.
-		// We will ignore any that are not set
-		log.WithValues("set states", stateReplicas).
-			WithValues("namespace state", state.Name).
-			Info("State could not be found")
-		return stateReplica, err
-	}
-	return stateReplica, err
-}
-
-func ObjectStateReplicasList(state states.State, deployments v1.DeploymentList) ([]sr.StateReplica, error) {
-
-	var stateReplicaList []sr.StateReplica
-	var err error
-
-	for _, deployment := range deployments.Items {
-		log := ctrl.Log.
-			WithValues("deployment", deployment.Name).
-			WithValues("namespace", deployment.Namespace)
-		stateReplicas, err := sr.NewStateReplicasFromAnnotations(deployment.GetAnnotations())
-		if err != nil {
-			log.WithValues("deployment", deployment.Name).
-				WithValues("namespace", deployment.Namespace).
-				Error(err, "Cannot calculate state replicas. Please check deployment annotations. Continuing.")
-			return []sr.StateReplica{}, err
-		}
-
-		stateReplica, err := stateReplicas.GetState(state.Name)
-		if err != nil {
-			// TODO here we should do priority filtering, and go down one level of priority to find the lowest set one.
-			// We will ignore any that are not set
-			log.WithValues("set states", stateReplicas).
-				WithValues("namespace state", state.Name).
-				Info("State could not be found")
-			return []sr.StateReplica{}, err
-		}
-
-		stateReplicaList = append(stateReplicaList, stateReplica)
-
-	}
-
-	return stateReplicaList, err
-}
-
 func ReconcileDeployment(ctx context.Context, _client client.Client, deployment v1.Deployment, state states.State, optIn bool) error {
 
 	log := ctrl.Log.
 		WithValues("deployment", deployment.Name).
 		WithValues("namespace", deployment.Namespace)
 
-	stateReplica, err := ObjectStateReplicas(state, deployment, optIn)
+	stateReplica, err := resources.DeploymentStateReplicas(state, deployment, optIn)
 	if err != nil {
 		log.Error(err, "Error getting the state replicas")
 		return err
@@ -182,7 +123,7 @@ func ReconcileDeployment(ctx context.Context, _client client.Client, deployment 
 	log.Info("Quota Check")
 
 	if allowed {
-		err = ScaleDeployment(ctx, _client, deployment, stateReplica)
+		err = resources.ScaleDeployment(ctx, _client, deployment, stateReplica)
 		if err != nil {
 			log.Error(err, "Error scaling the deployment")
 			return err
@@ -190,63 +131,6 @@ func ReconcileDeployment(ctx context.Context, _client client.Client, deployment 
 	}
 
 	return err
-}
-
-func ReconcileDeploymentforScalingState(ctx context.Context, _client client.Client, deployment v1.Deployment, stateReplica sr.StateReplica) error {
-
-	log := ctrl.Log.
-		WithValues("deployment", deployment.Name).
-		WithValues("namespace", deployment.Namespace)
-
-	err := ScaleDeployment(ctx, _client, deployment, stateReplica)
-	if err != nil {
-		log.Error(err, "Error scaling the deployment")
-		return err
-	}
-
-	return err
-}
-
-func ScaleDeployment(ctx context.Context, _client client.Client, deployment v1.Deployment, stateReplica sr.StateReplica) error {
-	log := ctrl.Log.
-		WithValues("deployment", deployment.Name).
-		WithValues("namespace", deployment.Namespace)
-
-	var err error
-	var oldReplicaCount int32
-	oldReplicaCount = *deployment.Spec.Replicas
-	retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if oldReplicaCount == stateReplica.Replicas {
-			log.Info("No Update on deployment. Desired replica count already matches current.")
-			return nil
-		}
-		log.Info("Updating deploymentconfig replicas for state", "replicas", stateReplica.Replicas)
-		updateErr := resources.DeploymentScaler(ctx, _client, deployment, stateReplica.Replicas)
-		if updateErr == nil {
-			log.WithValues("Deployment", deployment.Name).
-				WithValues("StateReplica mode", stateReplica.Name).
-				WithValues("Old Replica count", oldReplicaCount).
-				WithValues("New Replica count", stateReplica.Replicas).
-				Info("Deployment succesfully updated")
-			return nil
-		}
-		log.Info("Updating deployment failed due to a conflict! Retrying..")
-		// We need to get a newer version of the object from the client
-		var req reconcile.Request
-		req.NamespacedName.Namespace = deployment.Namespace
-		req.NamespacedName.Name = deployment.Name
-		deployment, err = resources.DeploymentGetter(ctx, _client, req)
-		if err != nil {
-			log.Error(err, "Error getting refreshed deployment in conflict resolution")
-		}
-		return updateErr
-
-	})
-	if retryErr != nil {
-		log.Error(retryErr, "Unable to scale the deployment, err: %v")
-	}
-
-	return nil
 }
 
 func ReconcileDeploymentConfig(ctx context.Context, _client client.Client, deploymentConfig ocv1.DeploymentConfig, state states.State, optIn bool) error {
